+++ applications/vessel-history/src/features/home/Home.tsx
@@ -1,14 +1,18 @@
-import React, { Fragment, useCallback, useEffect, useState } from 'react'
+import React, { Fragment, useEffect } from 'react'
 import cx from 'classnames'
 import { useDispatch, useSelector } from 'react-redux'
 import { DebounceInput } from 'react-debounce-input'
 import { useTranslation } from 'react-i18next'
 import Logo from '@globalfishingwatch/ui-components/dist/logo'
-import GFWAPI from '@globalfishingwatch/api-client'
-import { Spinner, IconButton } from '@globalfishingwatch/ui-components'
-import { VesselSearch } from '@globalfishingwatch/api-types'
-import { BASE_DATASET } from 'data/constants'
-import { getLastQuery, getVesselsFound, setVesselSearch } from 'features/search/search.slice'
+import { Spinner, IconButton, Button } from '@globalfishingwatch/ui-components'
+import { RESULTS_PER_PAGE } from 'data/constants'
+import {
+  getOffset,
+  getTotalResults,
+  getVesselsFound,
+  isSearching,
+  setOffset,
+} from 'features/search/search.slice'
 import { logoutUserThunk } from 'features/user/user.slice'
 import VesselListItem from 'features/vessel-list-item/VesselListItem'
 import { useOfflineVesselsAPI } from 'features/vessels/offline-vessels.hook'
@@ -16,7 +20,6 @@ import { selectAll as selectAllOfflineVessels } from 'features/vessels/offline-v
 import SearchPlaceholder, { SearchNoResultsState } from 'features/search/SearchPlaceholders'
 import { selectQueryParam } from 'routes/routes.selectors'
 import { useLocationConnect } from 'routes/routes.hook'
-import { OfflineVessel } from 'types/vessel'
 import styles from './Home.module.css'
 import '@globalfishingwatch/ui-components/dist/base.css'
 
@@ -31,55 +34,23 @@ interface LoaderProps {
 const Home: React.FC<LoaderProps> = (): React.ReactElement => {
   const { t } = useTranslation()
   const dispatch = useDispatch()
-  const [loading, setLoading] = useState(false)
-  const lastQuery = useSelector(getLastQuery)
+  const searching = useSelector(isSearching)
   const vessels = useSelector(getVesselsFound)
   const query = useSelector(selectQueryParam('q'))
+  const offset = useSelector(getOffset)
+  const totalResults = useSelector(getTotalResults)
   const offlineVessels = useSelector(selectAllOfflineVessels)
   const { dispatchQueryParams } = useLocationConnect()
   const { dispatchFetchOfflineVessels, dispatchDeleteOfflineVessel } = useOfflineVesselsAPI()
-
-  const minimumCharacters = 3
-  const resultsPerRequest = 25
 
   useEffect(() => {
     dispatchFetchOfflineVessels()
   }, [dispatchFetchOfflineVessels])
 
-  const fetchData = useCallback(
-    async (query: string) => {
-      setLoading(true)
-      GFWAPI.fetch<any>(
-        `/v1/vessels/search?datasets=${encodeURIComponent(
-          BASE_DATASET
-        )}&limit=${resultsPerRequest}&offset=${0}&query=${encodeURIComponent(query)}`
-      )
-        .then((json: any) => {
-          const resultVessels: Array<VesselSearch> = json.entries
-          setLoading(false)
-          dispatch(setVesselSearch({ vessels: resultVessels, query }))
-        })
-        .catch((error) => {
-          setLoading(false)
-        })
-    },
-    [dispatch]
-  )
-
-  useEffect(() => {
-    if (query?.length >= minimumCharacters && query !== lastQuery) {
-      fetchData(query)
-    }
-  }, [query, fetchData, lastQuery])
-
   const onInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     dispatchQueryParams({ q: e.target.value })
   }
-  const onDeleteClick = async (data: OfflineVessel) => {
-    setLoading(true)
-    await dispatchDeleteOfflineVessel(data.profileId)
-    setLoading(false)
-  }
+
   return (
     <div className={styles.homeContainer}>
       {!query && (
@@ -127,7 +98,7 @@ const Home: React.FC<LoaderProps> = (): React.ReactElement => {
                   key={index}
                   vessel={vessel}
                   saved={vessel.savedOn}
-                  onDeleteClick={() => onDeleteClick(vessel)}
+                  onDeleteClick={() => dispatchDeleteOfflineVessel(vessel.profileId)}
                 />
               ))}
             </div>
@@ -135,21 +106,39 @@ const Home: React.FC<LoaderProps> = (): React.ReactElement => {
         )}
         {query && (
           <Fragment>
-            <ul className={styles.searchResults}>
-              {loading && (
+            <div className={styles.searchResults}>
+              {searching && offset === 0 && (
                 <SearchPlaceholder>
                   <Spinner className={styles.loader}></Spinner>
                 </SearchPlaceholder>
               )}
-              {!loading && vessels.length > 0 && (
+              {(!searching || offset > 0) && vessels.length > 0 && (
                 <div className={styles.offlineVessels}>
                   {vessels.map((vessel, index) => (
                     <VesselListItem key={index} vessel={vessel} />
                   ))}
                 </div>
               )}
-              {!loading && vessels.length === 0 && <SearchNoResultsState />}
-            </ul>
+              {totalResults && !searching && vessels.length < totalResults && (
+                <div className={styles.listFooter}>
+                  <Button onClick={() => dispatch(setOffset(offset + RESULTS_PER_PAGE))}>
+                    {t('search.loadMore', 'LOAD MORE')}
+                  </Button>
+                </div>
+              )}
+              {searching && offset > 0 && (
+                <div className={styles.listFooter}>
+                  <Spinner className={styles.loader}></Spinner>
+                </div>
+              )}
+              {totalResults &&
+                !searching &&
+                vessels.length >= totalResults &&
+                vessels.length !== 0 && (
+                  <p className={styles.listFooter}>{t('search.noMore', 'NO MORE RESULTS')}</p>
+                )}
+              {!searching && vessels.length === 0 && <SearchNoResultsState />}
+            </div>
           </Fragment>
         )}
       </div>