+++ packages/fourwings-aggregate/src/index.ts
@@ -15,4 +15,4 @@ export {
   TileAggregationParams,
   TileAggregationSourceParams,
 } from './types'
-export { VALUE_MULTIPLIER } from './constants'
+export * from './constants'