+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.28.30
+
+### Patch Changes
+
+- Updated dependencies [fb321617]
+  - @globalfishingwatch/api-types@3.2.0
+
+## 0.28.29
+
+### Patch Changes
+
+- Updated dependencies [e2c17324]
+  - @globalfishingwatch/dataviews-client@6.2.3
+
+## 0.28.28
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/fourwings-aggregate@3.1.0
+  - @globalfishingwatch/layer-composer@5.2.3
+  - @globalfishingwatch/react-hooks@6.0.8
+  - @globalfishingwatch/ui-components@2.18.8
+
+## 0.28.27
+
+### Patch Changes
+
+- Updated dependencies [8d08df8a]
+  - @globalfishingwatch/layer-composer@5.2.2
+  - @globalfishingwatch/react-hooks@6.0.7
+  - @globalfishingwatch/ui-components@2.18.7
+
+## 0.28.26
+
+### Patch Changes
+
+- Updated dependencies [c0dc95ec]
+  - @globalfishingwatch/layer-composer@5.2.1
+  - @globalfishingwatch/react-hooks@6.0.6
+  - @globalfishingwatch/ui-components@2.18.6
+
 ## 0.28.25
 
 ### Patch Changes