+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @globalfishingwatchapp/vessel-history
 
+## 0.0.26
+
+### Patch Changes
+
+- @globalfishingwatch/layer-composer@5.2.3
+- @globalfishingwatch/react-hooks@6.0.8
+- @globalfishingwatch/ui-components@2.18.8
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [8d08df8a]
+  - @globalfishingwatch/layer-composer@5.2.2
+  - @globalfishingwatch/react-hooks@6.0.7
+  - @globalfishingwatch/ui-components@2.18.7
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [c0dc95ec]
+  - @globalfishingwatch/layer-composer@5.2.1
+  - @globalfishingwatch/react-hooks@6.0.6
+  - @globalfishingwatch/ui-components@2.18.6
+
 ## 0.0.23
 
 ### Patch Changes