+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @globalfishingwatch/ui-components
 
+## 2.18.8
+
+### Patch Changes
+
+- @globalfishingwatch/layer-composer@5.2.3
+
+## 2.18.7
+
+### Patch Changes
+
+- Updated dependencies [8d08df8a]
+  - @globalfishingwatch/layer-composer@5.2.2
+
+## 2.18.6
+
+### Patch Changes
+
+- Updated dependencies [c0dc95ec]
+  - @globalfishingwatch/layer-composer@5.2.1
+
 ## 2.18.5
 
 ### Patch Changes